chore(client): fix stale float annotations and stubs
Why: Clean up annotation drift and stubs from Decimal migration for audit finding [28] in docs/reviews/audit_2026-05-15.md
What:
- src/portfolio/tracker.py: update record_daily_snapshot and record_all_strategies signature to Decimal
- src/portfolio/summary.py: update type signatures to Decimal and remove float casts
- src/portfolio/formatting.py: update combined summary prices type to Decimal
- src/portfolio/service.py: update persist_snapshots parameters to Decimal
- src/paper/tracker.py: update prices type to Decimal and avoid float conversions
- src/dhan/reader.py: update enrich_with_upstox_prices signature to Decimal
- scripts/daily_snapshot.py: remove float casts on price and spot in historical snapshot loading
- scripts/paper_track_snapshot.py: update mock broker to return Decimal
- scripts/paper_3track_snapshot.py: update mock broker to return Decimal and add type hints
- scripts/paper_3track_overlay_roll.py: use quantize on Decimal spot prices instead of float cast
- tests/unit/dhan/test_reader.py: update test cases to pass Decimal prices
- tests/unit/paper/test_track_snapshot.py: update mock broker to return Decimal
- tests/unit/test_portfolio.py: update FakeMarket to convert input to Decimal and return Decimal dict
Ref: none

+++ src/portfolio/formatting.py
@@ -61,7 +61,7 @@ def _format_protection_stats(summary: PortfolioSummary) -> list[str]:
 
 def _format_combined_summary(
     strategies: list[Strategy],
-    prices: dict[str, float],
+    prices: dict[str, Decimal],
     strategy_pnls: dict[str, "StrategyPnL"],
     mf_pnl: "PortfolioPnL | None",
     prev_snapshots: dict[int, DailySnapshot] | None = None,
